requirements.txt
discord.py~=2.4
python-dotenv~=1.0
wordcloud~=1.9
Pillow~=10.4
numpy~=1.26